Lookup in library classes before program classes
For the internal tests where the build setup is broken (overlap
between library classes and program classes) the script for running
these tests will now first sanatize the library classes by removing
all classes from it which is also in the program classes (the program
classes are untouched).

Bug: 120884788

+++ tools/utils.py
@@ -337,16 +337,18 @@ def cloud_storage_exists(destination):
   return exit_code == 0
 
 class TempDir(object):
- def __init__(self, prefix=''):
+ def __init__(self, prefix='', delete=True):
    self._temp_dir = None
    self._prefix = prefix
+   self._delete = delete
 
  def __enter__(self):
    self._temp_dir = tempfile.mkdtemp(self._prefix)
    return self._temp_dir
 
  def __exit__(self, *_):
-   shutil.rmtree(self._temp_dir, ignore_errors=True)
+   if self._delete:
+     shutil.rmtree(self._temp_dir, ignore_errors=True)
 
 class ChangedWorkingDirectory(object):
  def __init__(self, working_directory, quiet=False):
